pgwire: move  maybeUpgradeToSecureConn to pre_serve.go
Release note: None
