+++ pkg/sql/pgwire/pre_serve.go
@@ -23,6 +23,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/hba"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
+	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgwirebase"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgwirecancel"
 	"github.com/cockroachdb/cockroach/pkg/util/metric"
 	"github.com/cockroachdb/cockroach/pkg/util/mon"
@@ -325,4 +326,84 @@ func (s *PreServeConnHandler) PreServe(
 
 	st.State = PreServeReady
 	return conn, st, nil
+}
+
+// maybeUpgradeToSecureConn upgrades the connection to TLS/SSL if
+// requested by the client, and available in the server configuration.
+func (s *PreServeConnHandler) maybeUpgradeToSecureConn(
+	ctx context.Context,
+	conn net.Conn,
+	connType hba.ConnType,
+	version uint32,
+	buf *pgwirebase.ReadBuffer,
+) (newConn net.Conn, newConnType hba.ConnType, newVersion uint32, clientErr, serverErr error) {
+	// By default, this is a no-op.
+	newConn = conn
+	newConnType = connType
+	newVersion = version
+	var n int // byte counts
+
+	if version != versionSSL {
+		// The client did not require a SSL connection.
+
+		// Insecure mode: nothing to say, nothing to do.
+		// TODO(knz): Remove this condition - see
+		// https://github.com/cockroachdb/cockroach/issues/53404
+		if s.cfg.Insecure {
+			return
+		}
+
+		// Secure mode: disallow if TCP and the user did not opt into
+		// non-TLS SQL conns.
+		if !s.cfg.AcceptSQLWithoutTLS && connType != hba.ConnLocal {
+			clientErr = pgerror.New(pgcode.ProtocolViolation, ErrSSLRequired)
+		}
+		return
+	}
+
+	if connType == hba.ConnLocal {
+		// No existing PostgreSQL driver ever tries to activate TLS over
+		// a unix socket. But in case someone, sometime, somewhere, makes
+		// that mistake, let them know that we don't want it.
+		clientErr = pgerror.New(pgcode.ProtocolViolation,
+			"cannot use SSL/TLS over local connections")
+		return
+	}
+
+	// Protocol sanity check.
+	if len(buf.Msg) > 0 {
+		serverErr = errors.Errorf("unexpected data after SSLRequest: %q", buf.Msg)
+		return
+	}
+
+	// The client has requested SSL. We're going to try and upgrade the
+	// connection to use TLS/SSL.
+
+	// Do we have a TLS configuration?
+	tlsConfig, serverErr := s.getTLSConfig()
+	if serverErr != nil {
+		return
+	}
+
+	if tlsConfig == nil {
+		// We don't have a TLS configuration available, so we can't honor
+		// the client's request.
+		n, serverErr = conn.Write(sslUnsupported)
+		if serverErr != nil {
+			return
+		}
+	} else {
+		// We have a TLS configuration. Upgrade the connection.
+		n, serverErr = conn.Write(sslSupported)
+		if serverErr != nil {
+			return
+		}
+		newConn = tls.Server(conn, tlsConfig)
+		newConnType = hba.ConnHostSSL
+	}
+	s.tenantIndependentMetrics.PreServeBytesOutCount.Inc(int64(n))
+
+	// Finally, re-read the version/command from the client.
+	newVersion, *buf, serverErr = s.readVersion(newConn)
+	return
 }
